Move backpack tests to its own file since it's not in entities anymore
