+++ tests/test_entities.py
@@ -115,48 +115,3 @@ class TestEntity(object):
         assert candy in game_map.entities
         candy.use(player)
         assert candy not in game_map.entities
-
-
-class TestBackpackItem(object):
-
-    def test_add_to_backpack(self, player, air):
-        player.backpack.add(air.key, 100)
-        assert air.key in player.backpack.contents
-        assert player.backpack.contents[air.key] == 100
-        assert player.backpack.cur_weight == 0.0
-
-    def test_use_unusable_item(self, player, air):
-        player.backpack.add(air.key, 1)
-        player.backpack.use(0, player)
-        assert player.backpack.contents[air.key] == 1
-
-    def test_add_existing_item(self, player, air):
-        player.backpack.add(air.key, 100)
-        player.backpack.add(air.key, 100)
-        assert len(player.backpack.contents) == 1
-        assert player.backpack.contents[air.key] == 200
-
-    def test_add_new_item(self, player, air, candy):
-        player.backpack.add(air.key, 100)
-        player.backpack.add(candy.key, 5)
-        assert len(player.backpack.contents) == 2
-
-    def test_add_over_limit_weight(self, player, air, candy):
-        player.backpack.add(candy.key, 100)
-        player.backpack.add(candy.key, 100)
-        player.backpack.add(air.key, 500)
-        assert player.backpack.contents[candy.key] == 100
-        assert player.backpack.contents[air.key] == 500
-        assert player.backpack.cur_weight == 100.0
-
-    def test_use_item(self, player, candy):
-        player.backpack.add(candy.key, 2)
-        player.hp -= 100
-        player.backpack.use(0, player)
-        assert player.backpack.contents[candy.key] == 1
-        assert player.hp == 52.5
-
-    def test_exhaust_item(self, player, candy):
-        player.backpack.add(candy.key, 1)
-        player.backpack.use(0, player)
-        assert len(player.backpack.contents) == 0
